Updated sigint handling
Updated sigint handling
Init on java client
Macro defs

+++ src/client/include/ctasks.h
@@ -9,7 +9,7 @@
 #include <shared_model.h>
 
 void *beat_loop(void *arg);
-void send_heartbeat(char *buffer, int buffSize, ConnData *connInfo);
+void send_heartbeat(char *buffer, size_t buffSize, ConnData *connInfo);
 char *make_packet(size_t dataSize, size_t frameSize, const char *header, const char *footer, size_t *packetSize);
 
 #endif
